requirements.txt
nornir==2.3.0
python-gnupg
hvac
